ArgLike (attempt 2)
Summary:
# Previous attempt

Previous attempt (D47861750) was reverted because it broke autodeps.

Previously I reverted the whole stack on unbreak autodeps quickly. This diff is likely not what causes autodeps breakage.

Context: autodeps uses `parser.py`, which is not exactly correct starlark parser.

# Original description

Replace `"_arglike"` with `ArgLike` constant.

```
ArgLike = typing.Any
```

can be more precise, but it might be costly.

Someone wrote `"_arglike"` instead of just `""` so I though it is important to preserve this information in the sources.

Reviewed By: ndmitchell

Differential Revision: D47913914

fbshipit-source-id: a550fe956c87baa56394e8bfe48c73a12af8f86c