Fix the problem that #compilation-database subtarget is broken for header-only library
Summary:
D46325905 changed the #compilation-database sub target behavior that it now generates an entry for the header file in the compilation database even if it belongs to a header-only library. This is not expected by the clang-tidy/codemode tool.

This changes the compilation database generation logic back so that it does not generate an entry for the header if it's a header only library.

At the same time, in the CxxCompilationDbInfo we still have an entry for the header by assuming it's C++ header, so C++ LS should not be affected.

CxxCompilationDbInfo is only used by C++ LS so no other service will be affected.

Note that this is a temporary solution until we deploy a new solution to find a source for header only target.

Reviewed By: milend

Differential Revision: D47960426

fbshipit-source-id: 4a7afb66266d8bb70e6d16abe5da77204ce2c8da