Revert D47988498: Multisect successfully blamed "D47988498: [buck2][haskell] Consider profiling when building artifacts" for test or build failures
Summary:
This diff is reverting D47988498
D47988498: [buck2][haskell] Consider profiling when building artifacts by gustavoavena has been identified to be causing the following test or build failures:

Tests affected:
- [sigma/haxl/tools/h2h/tests:golden - main](https://www.internalfb.com/intern/test/562950055417725/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2688729
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: gustavoavena

Differential Revision: D48105975

fbshipit-source-id: 36b297043dac503b04046eb1fac470730be8e557