Consider profiling when building artifacts (take #2)
Summary:
This is another version of D47988498, which the minor change that fixes the bug the broke a few of our tests.

The problem was that in the `haskell_ghci` rule, we were generating directory names using the suffix based on the profiling attribute of the `haskell_ghci` target, instead of the one considered when building the actual library.

This meant that the interface files were not found for some packages and their modules couldn't be imported.
---------
# Summary from D47988498

## This stack
See D47987891.

## This diff

Change the names of generated artifacts (e.g. package directories, lib names) to consider the status of profiling.

This is needed to (a) make it clear for every artifact if it was build with or without profiling and (b) to fix buck2 errors from declaring outputs with the same name of running the actions with the same category name (e.g. "haskell_link_...").

Reviewed By: helfper

Differential Revision: D48110784

fbshipit-source-id: e12ec14aacb406d2334bfcd22e5ba81b82c847eb