Record transitive external debug info in link info providers
Summary:
This is the part of {D47878969} factored out that **does not make any difference to compiler command-line arguments of any actions**, only additively affecting the set of files materialized during actions that involve debug info (link, dwp, dist LTO, BOLT), using `hidden`.

Thus it hopefully has no chance of breaking anything on its own. The breakage which led to the revert of the v2 diff linked above was related to the introduction of `-Csplit-debuginfo=unpacked` on the rustc command line, which causes debug info to be removed from rlibs and moved into external dwo files. That change is now in {D47928283}.

Reviewed By: zertosh

Differential Revision: D48174780

fbshipit-source-id: 12e8bfd188ebcbb3fbb05fee77242eb6a64f9e68