Universal: name universal outputs based off the sources
Summary:
**Problem**

`apple_universal_executable()` used its own name to for the name out of the output.

This meant that you need to set the name to the expected output, e.g., `libY.dylib` as the name of the target.

**Solution**

The expected behaviour would be for the binary to be named the same as the thin binary - for example, if we're building `libY.dylib`, create a `libY.dylib` universal dylib, etc.

Reviewed By: weih9

Differential Revision: D48158704

fbshipit-source-id: 799a500e82aacf9281c6bf4ba9a2f9124aa548e1