Pass full .jar classpath to d8 for non-root module invocations
Summary:
`buck1` does this here: https://www.internalfb.com/code/fbsource/[01cf4301875b]/xplat/build_infra/buck_client/src/com/facebook/buck/android/SmartDexingStep.java?lines=346%2C368-372

In `buck2`, we need to make sure that we provide a `classpath` to `d8` to make sure that it can desugar everything that it needs to.

The root module can't depend on anything outside of it, so don't bother with the classpath there.

For non-root modules, just pass every `.jar` to the `d8` invocation (except the `.jar` files that it is actually dex-ing).

Reviewed By: christolliday

Differential Revision: D48228252

fbshipit-source-id: 1f8a18fa0badc9c09f08f8c84870f23faf07d174