Create .aar with classes.jar only
Summary:
This is the beginning of an `android_aar` implementation - they are described here: https://developer.android.com/studio/projects/android-library#aar-contents

We get all the `.jar` files that we want to package into the `.aar`, and merge them into a single `classes.jar`. That can be easily accomplished using the `jar_builder` tool on the Java toolchain.

We then zip that entry up to create the `.aar` output.

Reviewed By: navidqar

Differential Revision: D48193336

fbshipit-source-id: dfb20f7f35f964e56bd2ab31ea43a55f0a73fa18