genrule: only make OUT relative to srcs in one place
Summary:
Currently $OUT for genrules is constructed as `"./{srcs_artifact}/../out"` and as long as `_requires_build_root(ctx)` is false (normal usage), it's made relative to `srcs_artifact` a second time here:
    289		  env_vars = {key: val.relative_to(srcs_artifact) for key, val in env_vars.items()}

If `requires_build_root` is false, `OUT="../out"`, but when it's true `OUT="buck-out/.../__genrule_target/srcs/../out/file"`.

This change removes the redundant `srcs/..` when `_requires_build_root=true`.

Having `srcs/..` in some Android targets that use `requires_build_root` (voltron module metadata) and that write output paths to a file causes failures in Sandcastle because the output path can't be resolved if srcs doesn't exist, and srcs doesn't exist if there is a cache hit on the genrule.

If needed this could be fixing by only changing behavior if `requires_build_root` is true, but this seems like it can be simplified since we are duplicating the logic used to create the output artifact path in `_declare_output`.

Reviewed By: cjhopman

Differential Revision: D48377228

fbshipit-source-id: 7565843772cd75c8aa23a2d187ed6ca0bee867dc