inplace: add support for main_function
Summary:
This diff introduces support for a new entry point in Python packages: a `main_function` that can be specified in addition to the `main_module`. This allows for more flexibility in defining the entry point of a Python package, and can be useful in certain cases where defining a module as the entry point is not feasible or desirable.

The diff also includes changes to the `make_py_package` tool to support this new entry point, as well as updates to the `python_binary` rule to allow for specifying both a `main_module` and a `main_function`.

Reviewed By: aleivag

Differential Revision: D48316132

fbshipit-source-id: 9881bebb3b401af879fd94f145cbecdd7713fc88