Use label.project_root to get the repo path for absolute argsfiles
Summary:
Currently, we pass a config (`xcode.absolute_path_prefix`) to get the project root in order to generate absolute paths for headermaps.

With recently added API, we can use `ctx.label.project_root` and `absolute=True` for write actions to create a file that has the user's repo path. This is then provided as an input to the `hmap_wrapper.py` tool to generate absolute path argsfiles.

Having to write a file per target is not ideal, but its a small locally written file and it gives us the caching/uniqueness we need for argsfile generation.

Reviewed By: milend, blackm00n

Differential Revision: D48448860

fbshipit-source-id: 688f641aaa8bbd3c6fad6a866e0e5f4818484864