Inspect Targets via BXL: Support "target_ref" alias target to locate rule providers
Summary:
## Context
Fdb currently works off materialized artifacts produced by "buck run --command-args-file". The problem is that in many cases it's not enough to properly setup debugger as different rule types might require additional setup and tweaks.
- for instance Java requires collecting class to source code mappings
- MacOS targets require building [debuginfo] to materialize symbols
- ARVR targets require getting unstripped libraries, and other environment setup
- Erlang requires passing specific arguments for VM
etc...
Additionally some rules produce trampoline-like scripts (Erlang) which are not reliable to parse and it'd be much better to instead inspect / extend rule providers to enrich with debugger related information

This has to live in "prelude" in order to be portable between repositories as I want to support Erlang debugging using this in the future.

Next series of diffs will make required setup to implement rules / targets inspection via BXL script

## This diff
Whenevery user invokes "fdb run <target>" the <target> is not necesserally mapped to a rule provider. This could some sort of wrapper as explained in D48517917. We need a way on a macro level to connect the dots for fdb to locate a target with rule provider that fdb supports.

This diff lays out the handling for special lable "fdb:rule:alias" that instructs Fdb to locate <target_name>fdb target which aliases to a rule provider that Fdb supports.

Reviewed By: splhack

Differential Revision: D48525455

fbshipit-source-id: 1abb812d247f3f0198471cd853e081e4fd6cc723