Allow running Nullsafe on pure Kotlin targets
Summary: Before this change building pure Kotlin `nullsafex-json` subtarget was failing due to the fact that no output was created (which makes sense since without Java sources Nullsafe analysis is not being run, therefore there is no output). To avoid confusion when the subtarget is available for some Kotlin targets but no to the others this output will be created by KotlinCD if it doesn't exist after running compilation process.

Reviewed By: IanChilds

Differential Revision: D48952509

fbshipit-source-id: bb3311eef0c8adbbfce3078f0041cafb6472ef84