Allow running Nullsafe on mixed Kotlin targets
Summary: Added `nullsafex-json` flavour to `kotlin_library`. The implementation is almost identical to what we've been doing in Java. `is_creating_subtarget` check is to prevent generating ABIs when running build with our subtarget. Otherwise, buck is not happy with some actions being bound multiple times (there was the same issue in Java integration which was solve the very same way).

Reviewed By: IanChilds

Differential Revision: D48952512

fbshipit-source-id: 1547bd82898f847a5945418458613651a10f17f9