Get rid of `extra_transitive_deps`
Summary: After the next diff, this became a bit finicky to handle - this diff replaces `extra_transitive_deps` with a simpler idea: Under the assumption that a rustc invocation's deps all compile, it should always be completely fine to pass `--extern` args to `-L` as well. So we can just do that for all doc tests. I don't think this would be wrong to do unconditionally, but is theoretically a waste of cycles

Reviewed By: shayne-fletcher

Differential Revision: D48847670

fbshipit-source-id: 43b1e3d5202571f3b9f09e23f877bb1e01f67ca3