Inspect via BXL: Traverse "alias" rules instead of falling back to default
Summary:
We base our decision how to inspect target on 2 things:
- rule.type
- labels (take precedense over rule.type)
It's quite common to have a chain of aliases e.g. "target1" -> "target1-android" (due to platform / os configurations) prior to obtain a rule or a target with "dbg:info:ref=another_target".

This diff introduces a fallback handling for "alias" and "configured_alias" rule types in the main flow.
Note that it used be handled for "dbg:info:ref" case separately now this works uniformly

Reviewed By: cjhopman

Differential Revision: D48855788

fbshipit-source-id: 52da67b94044e22993d4db7298427622feab4164