Add an is_buck2 function
Summary: I think we can probably optimise this a lot better than the existing thing in our macros, and it might one day be useful to test if you are Buck1 from the prelude. Note that currently if you test host_info from the prelude it is empty, so the usual Buck2 test can't be used. Thoughts?

Reviewed By: stepancheg

Differential Revision: D49053799

fbshipit-source-id: 771d1c8b7cbde12008d692f4592f7e5cba83a796