Introduce LibOutputStyle
Summary:
Currently, we use LinkStyle to represent two different concepts: (1)
the style of output for a library (e.g. (pic/non-pic) archive, shared lib) and
(2) the "link strategy" used when creating an executable/shared lib.

These aren't really directly connected to each other, except that they both
currently have three cases and that the default behavior for each strategy is
to pick the matching corresponding output style.

Using the same type and naming patterns for these makes it really confusing to
understand how things are used.

This change introduces "LibOutputStyle" to be used when the concept being encoded
is about the type of output of a particular library. I've updated most code to
use this as appropriate, but there was some code (particularly haskell+rust) where
it wasn't clear which concept this code was encoding.

For people familiar with this, really the main thing is that
`get_actual_link_style()` is changed to return a LibOutputStyle and
`create_linkable_node()`'s link_infos argument (the information about how to
include the node itself into a link output) is switched to be keyed on
LibOutputStyle. The other changes result from those two things.

I've used different case strings for this to make it further obvious that these two
enums aren't encoding the same thing, but where the case string ends up being exposed
to users, I've preserved the existing string (i think it'd be good to change that some
day, but it may be tricky).

This shouldn't change any behavior.

Reviewed By: chatura-atapattu

Differential Revision: D47808805

fbshipit-source-id: e6301a73af8eea4808fed8bfa82cf57d12cd697d