Pass in cxx toolchain rather than pulling it out of attrs
Summary:
In general, we shouldn't be accessing attrs in these shared functions
as it makes them harder to reuse and refactor. For cxx_toolchain, we've
already found that apple needs to extract it differently than c++. Android will
soon need to do it even more differently and this change will support that.

Reviewed By: stepancheg

Differential Revision: D49137194

fbshipit-source-id: c16b21222ece0b2ac48ae0d99bf7dc8fe171a27a