enforce that deps/exported_deps of a linkable node are consistent with the deps in the linkablegraph
Summary:
For linking using just the LinkableGraph, it is important that we not
traverse the entire LinkableGraph at each shared lib node. To support that,
we need to be able to discover the linkable node deps by just traversing
a single level of the graph children.

This change enforces that those are consistent in order to ensure that we
can traverse in that way.

Reviewed By: christolliday

Differential Revision: D49139111

fbshipit-source-id: c3ba5a18350bf65fe4df303d6761f65533974e8a