Add cxx_toolchain attr for android_binary
Summary:
For binary-level native things (somerge, relinker), we need a cxx
toolchain. In fact, we have a unique requirement that we actually have the
right toolchain for each of our cpu architectures and so it needs to get
the split transition that we apply for native deps.

Currently, buck doesn't have support for a split transition on a toolchain edge,
so we cheat and use a target edge. This means we don't actually get the exec
dep binding that we want. In practice this is okay for now as we'll get the fat
platform for both the binary and the toolchain node.

Reviewed By: christolliday

Differential Revision: D49141126

fbshipit-source-id: 1424301e19b77ba67a59325b9392869092e82fe8