enable allow_cache_upload for caffe2/torch/fb/training_toolkit/backend:worker
Summary:
Similar to D48149034, we want to make sure training_platform package is always
fast to build on warm revision. for that we need to make sure that binaries
that are built locally are uploaded from SC job to action cache, so no local rebuilds
are necessary (again, pretty much the same motivation as in D48149034). So this
diff does it for "knyfe" binary, and for libomnibus caffe2 worker.

Note that in parallel we should also migrate caffe2 backend worker to native
python

Reviewed By: Skory

Differential Revision: D49318078

fbshipit-source-id: 7bb3599ed4c68fa236609a73b3c81df777f673cb