Don't strip lib prefix from explicit sonames on Windows
Summary:
This fixes the Buck2 logic to not strip the "lib" prefix on Windows when it is explicitly specified in a `soname` argument for a C++ build target.

This was happening because the target formatting logic for handling `$(ext)` and `($ext  version)` substitution strings were implemented using a format string that included the prefix. Thus on Windows it was removed pre-emptively to not double-add it which worked in the case where it was originally being applied as the default, but not for the case where it was explicitly requested via parameter.

The fix is to split out the default prefix into its own `LinkerInfo` field and apply it in all cases except the explicit override path. This also made the versioned format string case a little more robust as it no longer implicitly requires that the format string for versioned extension on Windows only contain "dll".

Reviewed By: asp2insp

Differential Revision: D49398299

fbshipit-source-id: ce48c6e074663a8902cbbba375fc3fc036cf0442